change DataType For UserTime On TSheet

+++ TSheetProject/TSheet.Data/TimeSheetDetail.cs
@@ -11,7 +11,7 @@ namespace TSheet.Data
     {
         public int TimeSheetDetailID { get; set; }
 
-        public int? Hours { get; set; }
+        public Decimal? Hours { get; set; }
 
         public int TimeSheetMasterID { get; set; }
